# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Minha API",
      "version": "1.0.0",
      "description": "Descrição da minha API"
    },
    "paths": {
      "/usuarios": {
        "get": {
          "summary": "Obter a lista de usuários",
          "responses": {
            "200": {
              "description": "Lista de usuários retornada com sucesso"
            }
          }
        },
        "post": {
          "summary": "Criar um novo usuário",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Usuário criado com sucesso"
            },
            "400": {
              "description": "Dados inválidos fornecidos"
            }
          }
        }
      },
      "/usuarios/{id}": {
        "get": {
          "summary": "Obter informações de um usuário pelo ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Informações do usuário retornadas com sucesso"
            },
            "404": {
              "description": "Usuário não encontrado"
            }
          }
        },
        "put": {
          "summary": "Atualizar informações de um usuário pelo ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Informações do usuário atualizadas com sucesso"
            },
            "404": {
              "description": "Usuário não encontrado"
            }
          }
        },
        "delete": {
          "summary": "Excluir um usuário pelo ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Usuário excluído com sucesso"
            },
            "404": {
              "description": "Usuário não encontrado"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Usuario": {
          "type": "object",
          "properties": {
            "nome": {
              "type": "string"
            },
            "email": {
              "type": "string",
              "format": "email"
            }
          }
        }
      }
    }
  }
  
  
  